// packages/atomic/src/forms/types.ts
interface LocalizeFormOptionProps {
  checked?: React.InputHTMLAttributes<HTMLInputElement>['checked'];
  disabled?: React.InputHTMLAttributes<HTMLInputElement>['disabled'];
  required?: React.InputHTMLAttributes<HTMLInputElement>['required'];
  readOnly?: React.InputHTMLAttributes<HTMLInputElement>['readOnly'];
  error?: string;
}

export { LocalizeFormOptionProps };